# documentation/open-api.spec.yml
openapi: 3.1.0
info:
  title: Ecommerce
  description: |-
    This is the API documentation for our general purpose ecommerce REST API

    Some useful links:
    - [RESTful API repository](https://github.com/ilopez-ravn/ecommerce)


  contact:
    email: israellopez@ravn.co
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api/v1
  - url: https://ecommerce-api.pacasoft.app/api/v1
...
security:
  - bearerAuth: []  
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create a new user
      description: Create a new user
      operationId: createUser
      requestBody:
        description: User object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/login:
    post:
      tags:
        - users
      summary: Logs user into the system.
      description: Log user into the system.
      operationId: loginUser
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                properties:
                  access_token:
                    type: string

        '400':
          description: Invalid username/password supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                

        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - {}
  /users/refresh:
    post:
      tags:
        - users
      summary: Refresh JWT token
      description: Refresh the JWT token for an authenticated user
      operationId: refreshToken
      requestBody:
        description: Refresh token object
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
        required: true
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        '401':
          description: Unauthorized - Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        '400':
          description: Invalid refresh token supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - {}
  /users/password/token:
    post:
      tags:
        - users
      summary: Initiate password reset
      description: Initiate the password reset process for a user
      operationId: forgotPassword
      requestBody:
        description: User email for password reset
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email

        required: true
      responses:
        '200':
          description: Password reset initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset link sent to email
        '400':
          description: Invalid email supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
          
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - {}
  /users/password:
    put:
      tags:
        - users
      summary: Reset user password
      description: Reset the password for a user using a reset token
      operationId: resetPassword
      requestBody:
        description: Password reset data
        content:
          application/json:
            schema:
              type: object
              properties:
                reset_token:
                  type: string
                new_password:
                  type: string
        required: true
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password has been reset successfully
        '400':
          description: Invalid reset token or password supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - {}
  /users/{userId}:
    get:
      tags:
        - users
      summary: Get user by ID.
      description: Get user details base on ID.
      operationId: getUserByID
      parameters:
        - name: userId
          in: path
          description: The ID that needs to be fetched. Use 1 for testing
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
                
        '404':
          description: User not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put: 
      tags:
        - users
      summary: Update an existing user by ID
      description: Update an existing user by its unique identifier
      operationId: updateUserById
      parameters:
        - name: userId
          in: path
          description: ID of user to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: User object that needs to be updated
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role_id:
                  type: integer
                  format: int64
                phone:
                  type: string
                email:
                  type: string
                  format: email

                first_name:
                  type: string
                last_name:
                  type: string
        required: true
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                  username:
                    type: string
                  email:
                    type: string
                    format: email
                  first_name:
                    type: string
                  last_name:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"

        '404':
          description: User not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - users
      summary: Delete a user by ID
      description: Delete a single user by its unique identifier
      operationId: deleteUserById
      parameters:
        - name: userId
          in: path
          description: ID of user to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User deleted successfully
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"

        '404':
          description: User not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      


    patch:
      tags:
        - users
      summary: Disable or enable a user by ID
      description: Disable or enable a single user by its unique identifier
      operationId: disableOrEnableUserById
      parameters:
        - name: userId
          in: path
          description: ID of user to disable or enable
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Status to set for the user
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: boolean
        required: true

      responses:
        '200':
          description: User status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User status updated successfully
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"

        '404':
          description: User not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tags:
    get:
      tags:
        - tags
      summary: Get all product tags
      description: Retrieve a list of all product tags available
      operationId: getTags
      responses:
        '200':
          description: List of product tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - {}

    post:
      tags:
        - tags
      summary: Create a new product tag
      description: Create a new tag for products
      operationId: createTag
      requestBody:
        description: Tag object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTag'
        required: true
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tags/{tagId}:
    put:
      tags:
        - tags
      summary: Update an existing product tag
      description: Update an existing tag
      operationId: updateTag
      parameters:
        - name: tagId
          in: path
          description: ID of tag to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Tag object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTag'
        required: true
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"

        '404':
          description: Tag not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      

    delete:
      tags:
        - tags
      summary: Delete a product tag
      description: Delete a tag
      operationId: deleteTag
      parameters:
        - name: tagId
          in: path
          description: ID of tag to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Tag deleted successfully
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidQueryResponse"
        '404':
          description: Tag not found
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        first_name:
          type: string
          example: theEmployee
        last_name:
          type: string
          example: Jane
        email:
          type: string
          format: email
          example: Doe
        phone:
          type: string
          example: '12345'
        person_id: 
          type: integer
          format: int64
          example: 5
    CreateUser:
      type: object
      required: 
        - username
        - password
        - first_name
        - last_name
        - email
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: StrongPassword123
        role: 
          type: string
          example: manager
        first_name:
          type: string
          example: Israel
        last_name:
          type: string
          example: Lopez
        email:
          type: string
          format: email
          example: israellopez@ravn.co
        phone:
          type: string
          example: '12345'
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    NewTag:
      type: object
      required:
        - name
      properties:
        name:
          type: string
...
    Error:
      type: object
      properties:
        code:
          type: number
          example: 500
        message:
          type: string
          example: "Internal server error"
      required:
        - code
        - message
...
    InvalidQueryResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Invalid query parameter"
      required:
        - code
        - message
...
    UnauthorizedResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 401
        message:
          type: string
          example: "Unauthorized - Invalid or expired token"
      required:
        - code
        - message
  # requestBodies:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes